Feature/frontend auth

+++ css-cup-remastered-backend/src/main/java/ch/css/lernende/csscupremasteredbackend/service/TeamService.java
@@ -5,12 +5,11 @@ import ch.css.lernende.csscupremasteredbackend.dto.PlayerDto;
 import ch.css.lernende.csscupremasteredbackend.dto.TeamDto;
 import ch.css.lernende.csscupremasteredbackend.exception.IllegalParameterException;
 import ch.css.lernende.csscupremasteredbackend.exception.NoResultsFoundException;
-import ch.css.lernende.csscupremasteredbackend.model.Discipline;
-import ch.css.lernende.csscupremasteredbackend.model.Role;
-import ch.css.lernende.csscupremasteredbackend.model.TeamModel;
 import ch.css.lernende.csscupremasteredbackend.model.mapper.TeamMapper;
+import ch.css.lernende.csscupremasteredbackend.persistence.DisciplineEntity;
 import ch.css.lernende.csscupremasteredbackend.persistence.TeamEntity;
 import ch.css.lernende.csscupremasteredbackend.persistence.PlayerEntity;
+import ch.css.lernende.csscupremasteredbackend.repository.repo.discipline.DisciplineRepository;
 import ch.css.lernende.csscupremasteredbackend.repository.repo.player.PlayerRepository;
 import ch.css.lernende.csscupremasteredbackend.repository.repo.team.TeamRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -27,11 +26,13 @@ public class TeamService {
 
     private final TeamRepository teamRepository;
     private final PlayerRepository playerRepository;
+    private final DisciplineRepository disciplineRepository;
 
     @Autowired
-    public TeamService(TeamRepository teamRepository, PlayerRepository playerRepository) {
+    public TeamService(TeamRepository teamRepository, PlayerRepository playerRepository, DisciplineRepository disciplineRepository) {
         this.teamRepository = teamRepository;
         this.playerRepository = playerRepository;
+        this.disciplineRepository = disciplineRepository;
     }
 
     public List<TeamDto> fetchAll() {
@@ -58,36 +59,47 @@ public class TeamService {
     }
 
     public void addTeam(AddTeamDto addTeamDto) throws IllegalParameterException, SQLException {
-        //TODO: Remove teamp captain and implement real user
 
-        Optional<PlayerEntity> playerEntity = this.playerRepository.findById(addTeamDto.getUserId());
-        playerEntity.orElseThrow(IllegalParameterException::new);
+        Optional<DisciplineEntity> disciplineEntity = disciplineRepository.findByName(addTeamDto.getDiscipline().name());
 
-        teamRepository.insertTeam(addTeamDto.getName(), addTeamDto.getDiscipline(), playerEntity.get());
-        TeamEntity teamEntity = teamRepository.findByName(addTeamDto.getName());
+        if (disciplineEntity.isEmpty()) {
+            throw new IllegalParameterException();
+        }
 
-        for (PlayerDto entity : addTeamDto.getPlayers()) {
-            playerRepository.addTeamToPlayer(teamEntity, entity.getId());
+        teamRepository.insertTeam(addTeamDto.getTeamName(), disciplineEntity.get());
+        TeamEntity teamEntity = teamRepository.findByName(addTeamDto.getTeamName());
+
+        for (long id : addTeamDto.getPlayers()) {
+            Optional<PlayerEntity> playerEntity = playerRepository.findById(id);
+            playerRepository.addTeamToPlayer(teamEntity, playerEntity.orElseThrow(IllegalParameterException::new).getId());
         }
     }
 
-    public void deleteTeam(Optional<Long> id) throws IllegalArgumentException {
+    public void deleteTeam(Optional<Long> id) throws IllegalParameterException {
         if(id.isEmpty()) {
-            throw new IllegalArgumentException();
+            throw new IllegalParameterException();
         }
+
+        Optional<TeamEntity> teamEntity = teamRepository.findById(id.get());
+
+        if (teamEntity.isEmpty()) {
+            throw new IllegalParameterException();
+        }
+
+        teamEntity.get().getPlayers().stream().parallel().forEach(playerEntity -> {
+            playerEntity.setPlayerTeam(null);
+            playerRepository.save(playerEntity);
+        });
 
         teamRepository.deleteById(id.get());
     }
 
-    public void renameTeam(long id, String name) {
-        // TODO: Implement
-    }
+    public void joinTeam(Optional<Long> teamId, Optional<Long> userId) throws IllegalParameterException {
 
-    public void joinTeam(long userId, Optional<Long> id) throws IllegalParameterException {
+        long tId = teamId.orElseThrow(IllegalParameterException::new);
+        long uid = userId.orElseThrow(IllegalParameterException::new);
 
-        long teamId = id.orElseThrow(IllegalParameterException::new);
-
-        this.teamRepository.addPlayerToTeam(userId, teamId);
+        this.teamRepository.addPlayerToTeam(tId, uid);
 
     }
 }
